requirements.txt
flask
Pillow
torch
CLIP